+++ tests/Ozon.Route256.Five.OrderService.UnitTests/Features/GetOrderByIdQueryHandlerTest.cs
@@ -35,7 +35,7 @@ public class GetOrderByIdQueryHandlerTest
         var orderRepositoryMock = new Mock<IOrderRepository>();
         orderRepositoryMock.Setup(
                 x => x.GetOrderById(
-                    It.IsAny<int>(),
+                    It.IsAny<long>(),
                     It.IsAny<CancellationToken>()))
             .ReturnsAsync(orderData);
